add gigaset plugin
