+++ setup.py
@@ -25,6 +25,7 @@ setup(
             'api = wazo_phoned.plugins.api.plugin:Plugin',
             'aastra = wazo_phoned.plugins.aastra.plugin:Plugin',
             'cisco = wazo_phoned.plugins.cisco.plugin:Plugin',
+            'gigaset = wazo_phoned.plugins.gigaset.plugin:Plugin',
             'status = wazo_phoned.plugins.status.plugin:Plugin',
             'yealink = wazo_phoned.plugins.yealink.plugin:Plugin',
         ]
